+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [5.1.0] - Mixed-type import, unit-class
+
+### Added
+
+#### Mixed-Type Import
+
+- **Mixed-type import support**: A single CSV/TSV/JSON file can now contain both measurement statistics (min/max/mean) and counter statistics (sum/state)
+- **Consistent Unit checks**: If an import file contains different units for the same statistic id, there is an error
+
+Note:
+
+- Delta column remains incompatible with mixed imports — a file with a `delta` column cannot also contain `min`, `max`, `mean`, `sum`, or `state` columns
+
+#### External statistics can be used in the energy board
+
+- When **importing external statistics**, the unit-class (e.g. energy, power, ...) is set automatically depending on the imported unit. This allows to use external statistics with e.g. unit kWh to use them in the energy board.
+
 ## [5.0.2]
 
 ### Bug Fixes